implemented rating mvc controller

+++ VirtualTeacher/Models/DTOs/Course/RatingCreateDto.cs
@@ -4,7 +4,7 @@ namespace VirtualTeacher.Models.DTOs.Course;
 
 public class RatingCreateDto
 {
-    public byte Value { get; set; }
+    public double Value { get; set; }
     [MinLength(5, ErrorMessage = "The rating review must be at least 5 characters long.")]
     [MaxLength(70, ErrorMessage = "The rating review must be less than 70 characters long.")]
     public string? Review { get; set; }
