+++ VirtualTeacher/Models/DTOs/Course/RatingResponseDto.cs
@@ -5,7 +5,7 @@ namespace VirtualTeacher.Models.DTOs.Course;
 
 public class RatingResponseDto
 {
-    public byte Rating { get; set; }
+    public double Rating { get; set; }
     public string? Review { get; set; }
     public string Student { get; set; } = null!;
 }
